feat: enhance SummarizeTextAnalysisTask with llm_conf parameter and update tests
